requirements.txt
PyYAML
requests
tenacity